Thread log manager through resource init context
