+++ python_modules/dagster/dagster/core/init_context.py
@@ -4,7 +4,7 @@ import warnings
 
 from dagster import check
 
-from .definitions.logger import LoggerDefinition
+from ..core.log_manager import DagsterLogManager
 from .definitions.pipeline import PipelineDefinition
 from .definitions.resource import ResourceDefinition
 
@@ -45,7 +45,7 @@ class InitContext(namedtuple('_InitContext', 'context_config pipeline_def run_id
 
 class InitResourceContext(
     namedtuple(
-        'InitResourceContext', 'context_config resource_config pipeline_def resource_def run_id'
+        'InitResourceContext', 'context_config resource_config pipeline_def resource_def run_id log'
     )
 ):
     '''
@@ -63,9 +63,10 @@ class InitResourceContext(
         pipeline_def (PipelineDefinition): The pipeline definition currently being executed.
         resource_def (ResourceDefinition): The resource definition for the resource being constructed.
         run_id (str): The ID for this run of the pipeline.
+        log (DagsterLogManager): THe log manager for this run of the pipeline
     '''
 
-    def __new__(cls, context_config, resource_config, pipeline_def, resource_def, run_id):
+    def __new__(cls, context_config, resource_config, pipeline_def, resource_def, run_id, log=None):
         return super(InitResourceContext, cls).__new__(
             cls,
             context_config,
@@ -73,6 +74,7 @@ class InitResourceContext(
             check.inst_param(pipeline_def, 'pipeline_def', PipelineDefinition),
             check.inst_param(resource_def, 'resource_def', ResourceDefinition),
             check.str_param(run_id, 'run_id'),
+            check.opt_inst_param(log, 'log', DagsterLogManager),
         )
 
     @property
@@ -84,33 +86,3 @@ class InitResourceContext(
             )
         )
         return self.resource_config
-
-
-class InitLoggerContext(
-    namedtuple('InitLoggerContext', 'context_config logger_config pipeline_def logger_def run_id')
-):
-    '''
-    Similar to InitContext, but is logger-specific. It includes all the properties
-    in the InitContext, plus the logger config and the logger definition.
-
-
-    Attributes:
-        context_config (Any): The configuration data provided by the environment config. The
-            schema for this data is defined by ``config_field`` on the
-            :py:class:`PipelineContextDefinition`
-        logger_config (Any): The configuration data provided by the environment config. The
-            schema for this data is defined by ``config_field`` on the :py:class:`LoggerDefinition`
-        pipeline_def (PipelineDefinition): The pipeline definition currently being executed.
-        logger_def (LoggerDefinition): The logger definition for the logger being constructed.
-        run_id (str): The ID for this run of the pipeline.
-    '''
-
-    def __new__(cls, context_config, logger_config, pipeline_def, logger_def, run_id):
-        return super(InitLoggerContext, cls).__new__(
-            cls,
-            context_config,
-            logger_config,
-            check.inst_param(pipeline_def, 'pipeline_def', PipelineDefinition),
-            check.inst_param(logger_def, 'logger_def', LoggerDefinition),
-            check.str_param(run_id, 'run_id'),
-        )
