+++ python_modules/dagster/dagster_tests/utils_tests/log_tests/test_json_logging.py
@@ -4,7 +4,7 @@ import logging
 from contextlib import contextmanager
 
 from dagster import PipelineDefinition
-from dagster.core.init_context import InitLoggerContext
+from dagster.core.init_logger_context import InitLoggerContext
 from dagster.utils.log import define_json_file_logger
 from dagster.utils.test import create_test_pipeline_execution_context, get_temp_file_name
